third_party/blink/web_tests/virtual/css-anchor-update-enabled/external/wpt/css/css-anchor-position/anchor-scroll-position-try-013-expected.txt

This is a testharness.js-based test.
[FAIL] anchor-scroll-position-try-013 2
  assert_equals: Anchored element should be at the top of anchor expected 119 but got 219
Harness: the test ran to completion.
